+++ opennms-webapp/src/main/java/org/opennms/web/controller/alarm/AlarmFilterController.java
@@ -164,7 +164,7 @@ public class AlarmFilterController extends AbstractController implements Initial
         OnmsAlarm[] alarms = m_webAlarmRepository.getMatchingAlarms(AlarmUtil.getOnmsCriteria(queryCriteria));
         
         // get the total alarm count
-        int alarmCount = m_webAlarmRepository.countMatchingAlarms(AlarmUtil.getOnmsCriteria(countCriteria));
+        long alarmCount = m_webAlarmRepository.countMatchingAlarms(AlarmUtil.getOnmsCriteria(countCriteria));
         
         ModelAndView modelAndView = new ModelAndView(getSuccessView());
         modelAndView.addObject("alarms", alarms);
