+++ opennms-webapp/src/main/java/org/opennms/web/rest/AcknowledgmentRestService.java
@@ -116,7 +116,7 @@ public class AcknowledgmentRestService extends OnmsRestService {
     public String getCount() {
         readLock();
         try {
-            return Integer.toString(m_ackDao.countAll());
+            return Long.toString(m_ackDao.countAll());
         } finally {
             readUnlock();
         }
