+++ opennms-webapp/src/main/java/org/opennms/web/rest/AlarmRestService.java
@@ -112,7 +112,7 @@ public class AlarmRestService extends AlarmRestServiceBase {
     public String getCount() {
         readLock();
         try {
-            return Integer.toString(m_alarmDao.countAll());
+            return Long.toString(m_alarmDao.countAll());
         } finally {
             readUnlock();
         }
